Remove gdc2xena as an entry point
