+++ tests/test_main.py
@@ -30,3 +30,50 @@ class ParserTest(unittest.TestCase):
         assert parsed.outdir == "path/to/dir"
         assert parsed.name == "new_name"
         assert parsed.cohort == "cohort_name"
+
+    def test_etl(self):
+        parsed = self.parser.parse_args([
+            "etl",
+            "-r",
+            "path/to/dir",
+            "-p",
+            "project_name",
+            "-t",
+            "datatype",
+        ])
+        assert parsed.subcomm == "etl"
+        assert parsed.root == "path/to/dir"
+        assert parsed.projects == ["project_name"]
+        assert parsed.datatype == ["datatype"]
+        # for mutually exclusive groups
+        parsed = self.parser.parse_args([
+            "etl",
+            "-r",
+            "path/to/dir",
+            "-P",
+            "not_this_project_name",
+            "-T",
+            "not_this_datatype",
+        ])
+        assert parsed.subcomm == "etl"
+        assert parsed.root == "path/to/dir"
+        assert parsed.not_projects == ["not_this_project_name"]
+        assert parsed.not_datatype == ["not_this_datatype"]
+
+    def test_metaparser(self):
+        parsed = self.parser.parse_args([
+            "metadata",
+            "-p",
+            "project_name",
+            "-t",
+            "datatype",
+            "-m",
+            "path/to/matrix",
+            "-r",
+            "10",
+        ])
+        assert parsed.subcomm == "metadata"
+        assert parsed.project == "project_name"
+        assert parsed.datatype == "datatype"
+        assert parsed.matrix == "path/to/matrix"
+        assert parsed.release == 10.0
