+++ xena_gdc_etl/gdc2xena.py
@@ -92,109 +92,3 @@ def gdc2xena(root_dir, projects, xena_dtypes):
     m, s = divmod(int(end_time - start_time), 60)
     h, m = divmod(m, 60)
     print('Finish in {:d}:{:02d}:{:02d}.'.format(h, m, s))
-
-
-def create_parser():
-    parser = argparse.ArgumentParser(
-            description='Pipeline for importing data from GDC to Xena.'
-        )
-    subparsers = parser.add_subparsers(title='Subcommands', dest='subcomm',
-                                       metavar='')
-    # Subcommand for full ETL (download, transform, and metadata)
-    etlparser = subparsers.add_parser(
-            'etl',
-            help='Download and transform GDC data into Xena matrix, '
-                 'and generate corresponding metadata.',
-            epilog='Supported data types are: {}'.format(str(valid_dtype))
-        )
-    etlparser.add_argument('-r', '--root', type=str, default='.',
-                           help='Root directory for imported data.')
-    projects_group = etlparser.add_mutually_exclusive_group()
-    projects_group.add_argument('-p', '--projects', type=str, nargs='+',
-                                help='GDC project ID(s) to be imported; or '
-                                     '"all" if all projects on GDC are going '
-                                     'to be imported. Defaults to "all".',
-                                default=['all'])
-    projects_group.add_argument('-P', '--not-projects', type=str, nargs='+',
-                                help='Import all projects on GDC except '
-                                     'projects specified by this option. '
-                                     'This option and the "-p" option are '
-                                     'mutually exclusive.',
-                                default=[])
-    datatype_group = etlparser.add_mutually_exclusive_group()
-    datatype_group.add_argument('-t', '--datatype', type=str, nargs='+',
-                                help='Data type code(s) to be imported; or '
-                                     '"all" if all supported types are going '
-                                     'to be imported. Defaults to "all".',
-                                default=['all'])
-    datatype_group.add_argument('-T', '--not-datatype', type=str, nargs='+',
-                                help='Import all supported types except '
-                                     'projects specified by this option. '
-                                     'This option and the "-t" option are '
-                                     'mutually exclusive.',
-                                default=[])
-
-    # Subcommand for making metadata
-    metaparser = subparsers.add_parser(
-            'metadata',
-            help='Generate metadata for a Xena matrix',
-            epilog='Supported data types are: {}'.format(str(valid_dtype))
-        )
-    metaparser.add_argument('-p', '--project', type=str, required=True,
-                            help='The project of the matrix.')
-    metaparser.add_argument('-t', '--datatype', type=str, required=True,
-                            help='One data type code for the matrix.')
-    metaparser.add_argument('-m', '--matrix', type=str, required=True,
-                            help='Path to a Xena matrix')
-    metaparser.add_argument('-r', '--release', type=float, required=True,
-                            help='GDC data release number.')
-    return parser
-
-
-def main():
-    """
-    Program entry point for gdc2xena
-    """
-    parser = create_parser()
-    args = parser.parse_args()
-    if args.subcomm == 'etl':
-        root_dir = os.path.abspath(args.root)
-        projects = args.projects
-        if 'all' in [p.lower() for p in projects]:
-            projects = [str(x) for x in gdc.get_project_info().index]
-        for p in args.not_projects:
-            projects.remove(p)
-        xena_dtypes = args.datatype
-        if 'all' in [t.lower() for t in xena_dtypes]:
-            xena_dtypes = valid_dtype
-        for t in args.not_datatype:
-            xena_dtypes.remove(t)
-        print('#### GDC to Xena Importing Settings ####')
-        total_projects = len(projects)
-        print('Import the following {} projects:'.format(total_projects))
-        print(repr(projects), end='\n\n')
-        print('for the following {} types of data:'.format(len(xena_dtypes)))
-        print(str(xena_dtypes), end='\n\n')
-        print('into this directory: {}'.format(root_dir))
-        print('########################################', end='\n\n')
-        gdc2xena(root_dir, projects, xena_dtypes)
-    elif args.subcomm == 'metadata':
-        root_dir = os.path.dirname(args.matrix)
-        if args.datatype == 'survival':
-            dataset = GDCSurvivalset(args.project, root_dir)
-        elif args.datatype == 'raw_phenotype':
-            if args.project.startswith('TCGA'):
-                dataset = GDCPhenoset(args.project, 'raw_phenotype',
-                                      root_dir)
-            if args.project.startswith('TARGET'):
-                dataset = GDCPhenoset(args.project, 'clinical', root_dir)
-        elif args.datatype == 'GDC_phenotype':
-            dataset = GDCPhenoset(args.project, 'GDC_phenotype', root_dir)
-        else:
-            dataset = GDCOmicset(args.project, args.datatype, root_dir)
-        dataset.matrix = args.matrix
-        dataset.gdc_release = (
-                'https://docs.gdc.cancer.gov/Data/Release_Notes/Data_Release_Notes/#data-release-' +  # noqa
-                str(args.release).replace('.', '')
-            )
-        dataset.metadata()
